refactoring opts with arguments and files
